Move set_key_rotation_policy into application code
