+++ src/repositories/Maps.ts
@@ -13,6 +13,6 @@ export const keyReleaseMapName = "public:ccf.gov.policies.key_release";
 export const keyReleasePolicyMap = ccf.kv[keyReleaseMapName];
 export const settingsMapName = "public:policies.settings";
 export const settingsPolicyMap = ccf.kv[settingsMapName];
-export const keyRotationMapName = "public:ccf.gov.policies.key_rotation";
+export const keyRotationMapName = "public:policies.key_rotation";
 export const keyRotationPolicyMap = ccf.kv[keyRotationMapName];
 //#endregion
